Extract Eq for Eval, Function0 and IO
