+++ kategory/src/main/kotlin/kategory/instances/EvalEq.kt
@@ -1,0 +1,6 @@
+package kategory
+
+class EvalEq : Eq<HK<Eval.F, Int>> {
+    override fun eqv(a: HK<Eval.F, Int>, b: HK<Eval.F, Int>): Boolean =
+            a.ev().value() == b.ev().value()
+}
