+++ kategory/src/main/kotlin/kategory/instances/Function0Eq.kt
@@ -1,0 +1,6 @@
+package kategory
+
+class Function0Eq : Eq<HK<Function0.F, Int>> {
+    override fun eqv(a: HK<Function0.F, Int>, b: HK<Function0.F, Int>): Boolean =
+            a.ev()() == b.ev()()
+}
