+++ kategory/src/test/kotlin/kategory/data/Function0Test.kt
@@ -8,10 +8,7 @@ import org.junit.runner.RunWith
 class Function0Test : UnitSpec() {
     init {
 
-        testLaws(MonadLaws.laws(Function0, object : Eq<HK<Function0.F, Int>> {
-            override fun eqv(a: HK<Function0.F, Int>, b: HK<Function0.F, Int>): Boolean =
-                    a.ev()() == b.ev()()
-        }))
+        testLaws(MonadLaws.laws(Function0, Function0Eq()))
 
         "Function0Monad.binding should for comprehend over all values of multiple Function0" {
             Function0.binding {
