+++ kategory/src/test/kotlin/kategory/effects/IOTest.kt
@@ -11,10 +11,7 @@ class IOTest : UnitSpec() {
 
     init {
 
-        testLaws(MonadLaws.laws(IO, object : Eq<HK<IO.F, Int>> {
-            override fun eqv(a: HK<IO.F, Int>, b: HK<IO.F, Int>): Boolean =
-                a.ev().unsafeRunSync() == b.ev().unsafeRunSync()
-        }))
+        testLaws(MonadLaws.laws(IO, IOEq()))
 
         "should defer evaluation until run" {
             var run = false
